#129 메모 리펙토링
