+++ EUMMEYO.xcodeproj/xcuserdata/evan-kim.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -62,8 +62,8 @@
             filePath = "EUMMEYO/View/Calendar/CalendarView.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "520"
-            endingLineNumber = "520"
+            startingLineNumber = "521"
+            endingLineNumber = "521"
             landmarkName = "DayView(day:)"
             landmarkType = "7">
          </BreakpointContent>
@@ -84,28 +84,28 @@
             landmarkType = "14">
             <Locations>
                <Location
-                  uuid = "EED955E9-6AA8-4D33-AE94-3A3AED697827 - 9146c9eb9f79895e"
+                  uuid = "EED955E9-6AA8-4D33-AE94-3A3AED697827 - b72b75df3394641b"
                   shouldBeEnabled = "Yes"
                   ignoreCount = "0"
                   continueAfterRunningActions = "No"
                   symbolName = "property wrapper backing initializer of EUMMEYO.NicknameSettingView.birthdate : Foundation.Date"
                   moduleName = "EUMMEYO.debug.dylib"
                   usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/evan-kim/Documents/Development/Team%20Eummeyo/250412/EUMMEYO/EUMMEYO/View/Nickname/NicknameSettingView.swift"
+                  urlString = "file:///Users/evan-kim/Documents/Development/Team%20Eummeyo/250507/EUMMEYO/EUMMEYO/View/Nickname/NicknameSettingView.swift"
                   startingColumnNumber = "9223372036854775807"
                   endingColumnNumber = "9223372036854775807"
                   startingLineNumber = "16"
                   endingLineNumber = "16">
                </Location>
                <Location
-                  uuid = "EED955E9-6AA8-4D33-AE94-3A3AED697827 - ae99614a2688ba0a"
+                  uuid = "EED955E9-6AA8-4D33-AE94-3A3AED697827 - 88f4dd7e8a65574f"
                   shouldBeEnabled = "Yes"
                   ignoreCount = "0"
                   continueAfterRunningActions = "No"
                   symbolName = "EUMMEYO.NicknameSettingView.init(authViewModel: SwiftUI.EnvironmentObject&lt;EUMMEYO.AuthenticationViewModel&gt;) -&gt; EUMMEYO.NicknameSettingView"
                   moduleName = "EUMMEYO.debug.dylib"
                   usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/evan-kim/Documents/Development/Team%20Eummeyo/250412/EUMMEYO/EUMMEYO/View/Nickname/NicknameSettingView.swift"
+                  urlString = "file:///Users/evan-kim/Documents/Development/Team%20Eummeyo/250507/EUMMEYO/EUMMEYO/View/Nickname/NicknameSettingView.swift"
                   startingColumnNumber = "9223372036854775807"
                   endingColumnNumber = "9223372036854775807"
                   startingLineNumber = "16"
