--- EUMMEYO/Device/Audio.swift
@@ -1,8 +1,0 @@
-//
-//  Audio.swift
-//  EUMMEYO
-//
-//  Created by eunchanKim on 4/30/25.
-//
-
-import Foundation
