--- EUMMEYO/Service/AudioPlayerService.swift
@@ -1,8 +1,0 @@
-//
-//  AudioPlayerService.swift
-//  EUMMEYO
-//
-//  Created by eunchanKim on 4/29/25.
-//
-
-import Foundation
