+++ EUMMEYO/View/Calendar/CalendarView.swift
@@ -214,7 +214,8 @@ struct CalendarView: View {
                 } else {
                     ForEach(memos){ memo in
                         NavigationLink {
-                            MemoDetailView(memo: memo, editMemo: memo.content, editTitle: memo.title)
+                            MemoDetailView(memo: memo, container: container)
+                                .environmentObject(calendarViewModel)
                         } label: {
                             MemoCardView(memo: memo)
                         }
@@ -526,7 +527,7 @@ struct CalendarView: View {
 }
 
 struct MemoCardView: View {
-    var memo: Memo
+    @State var memo: Memo
     @EnvironmentObject var viewModel: CalendarViewModel
     @State var offsetX: CGFloat = 0 // 드래그 거리
     @State var showDelete: Bool = false // 삭제 버튼 표시 여부
@@ -599,8 +600,8 @@ struct MemoCardView: View {
                         Text(memo.date.formatted(date: .numeric, time: .omitted))
                             .font(.system(size: 15.scaled))
                         Button {
-                            viewModel.isBookmark.toggle()
-                            viewModel.toggleBookmark(memoId: memo.id, isBookmark: viewModel.isBookmark)
+                            memo.isBookmarked.toggle()
+                            viewModel.toggleBookmark(memoId: memo.id, isBookmark:  memo.isBookmarked)
                         } label: {
                             Image(systemName: memo.isBookmarked ? "star.fill" : "star")
                                 .foregroundColor(memo.isBookmarked ? .mainPink : .mainGray)
